# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sarjy API",
    "description": "\nBackend for **Sarjy**, a voice assistant that remembers things across sessions.\n\nSpeech-to-text and text-to-speech both run in the browser via the Web Speech\nAPI, so this API only ever handles text.\n\n### Identity\n\nThere is no authentication. `user_id` is a UUID generated by the browser and\nkept in `localStorage`; the server accepts it as given. Reads are scoped by it\nand return `404` for another user's session, which prevents accidental\ncross-user reads rather than determined ones. This is a deliberate tradeoff for\na demo, recorded in the project's PRD.\n\n### Sessions\n\nA session row is created lazily, on its first message, so empty conversations\nnever reach the database. Deleting a session removes its messages but preserves\nany memories learned in it \u2014 durable facts outlive the conversation that taught\nthem.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send a message and get Sarjy's reply",
        "description": "Run one conversational turn.\n\nCreates the user and session rows if they are new, persists the incoming\nmessage, replays the recent history along with everything Sarjy remembers\nabout this user, and returns the reply.\n\nThe model may call `save_memory` during the turn to record a durable fact.\nThat write is committed as part of the same turn and is visible to every\nlater session for this user.\n\nOutbound model calls pass through a token-bucket limiter that makes callers\nwait rather than rejecting them, so a busy server shows up as latency.",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sarjy's reply text.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "A path or query id is not a UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "The local token bucket would have queued this turn past its cap.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The model could not be reached, even after retries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send a message and stream Sarjy's reply",
        "description": "Run one conversational turn, streamed.\n\nIdentical to `POST /api/chat` in what it persists; the difference is that\ntokens are emitted as they arrive, so the browser can begin speaking at the\nfirst sentence boundary rather than at the last token. Frames are\n`data: {json}` and carry `type` of `delta`, `done`, or `error`; failures\narrive as an `error` frame with HTTP 200, since the response has usually\nalready begun by then.",
        "operationId": "chat_stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A `text/event-stream` of `delta` frames, then one `done` frame.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "A path or query id is not a UUID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Client-generated UUID identifying the browser."
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Client-generated UUID for the conversation. The row is created on the first message, so a session the user has opened but not written in does not exist server-side yet."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "The user's turn, typed or transcribed."
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "session_id",
          "message"
        ],
        "title": "ChatRequest",
        "example": {
          "message": "What's the weather like?",
          "session_id": "b74e2f10-8c3a-4d61-9f52-1a7e0c9d4b88",
          "user_id": "3f1a9c2e-5b7d-4e88-9a21-6c0f4d8b1e33"
        }
      },
      "ChatResponse": {
        "properties": {
          "reply": {
            "type": "string",
            "title": "Reply",
            "description": "Sarjy's reply. Spoken aloud by the client unless muted."
          },
          "timings": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timings",
            "description": "Server-side spans for this turn, in milliseconds: `db_read_ms`, `db_write_ms`, `limiter_wait_ms`, `llm_total_ms`, `llm_ttft_ms` (null unless streamed) and `total_ms`. Present so the client can attribute latency without server access."
          }
        },
        "type": "object",
        "required": [
          "reply"
        ],
        "title": "ChatResponse",
        "example": {
          "reply": "It's 34\u00b0C and clear in Riyadh right now."
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "example": {
          "detail": "user_id must be a UUID"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}